backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
celery==5.5.3
redis==6.4.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
PyMuPDF==1.26.4
elasticsearch==8.19.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
pydantic==2.11.7
pytest==8.4.1
